[MOD]opd-4878 修改最近1年营业总额、最近2年营业总额由原先的Long类型修改为BigDecimal

See merge request op-deliver-rcwl-group/srm-source-op!469

+++ src/main/java/org/srm/source/cux/domain/entity/RcwlShortlistHeader.java
@@ -8,6 +8,7 @@ import javax.validation.constraints.NotNull;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -89,9 +90,9 @@ public class RcwlShortlistHeader extends ExpandDomain {
     @ApiModelProperty(value = "成立年限")
     private Long years;
     @ApiModelProperty(value = "近1年营业额（万元）")
-    private Long oneProfit;
+    private BigDecimal oneProfit;
     @ApiModelProperty(value = "近2年营业额（万元）")
-    private Long twoProfit;
+    private BigDecimal twoProfit;
     @ApiModelProperty(value = "状态")
     private String state;
     @ApiModelProperty(value = "是否审核")
@@ -341,22 +342,22 @@ public class RcwlShortlistHeader extends ExpandDomain {
     /**
      * @return 近1年营业额（万元）
      */
-    public Long getOneProfit() {
+    public BigDecimal getOneProfit() {
         return oneProfit;
     }
 
-    public void setOneProfit(Long oneProfit) {
+    public void setOneProfit(BigDecimal oneProfit) {
         this.oneProfit = oneProfit;
     }
 
     /**
      * @return 近2年营业额（万元）
      */
-    public Long getTwoProfit() {
+    public BigDecimal getTwoProfit() {
         return twoProfit;
     }
 
-    public void setTwoProfit(Long twoProfit) {
+    public void setTwoProfit(BigDecimal twoProfit) {
         this.twoProfit = twoProfit;
     }
